Added reading cities from file

+++ src/Evolution.hpp
@@ -4,11 +4,12 @@
 #include <vector>
 #include "Problem.hpp"
 #include "Result.hpp"
+#include "FileReader.hpp"
 
 class Evolution
 {
 public:
-	Evolution(int pop_size, int gen, int px, int pm, int tour, std::string filename);
+	Evolution(int pop_size, int gen, float px, float pm, int tour, std::string filename);
 	void step();
 
 private:
